Split loader file into multiple different files
