+++ python_modules/dagster/dagster_tests/definitions_tests/module_loader_tests/checks_module/checks_submodule_2/__init__.py
@@ -1,3 +1,3 @@
-from dagster_tests.definitions_tests.asset_check_tests.checks_module.checks_submodule import (
+from dagster_tests.definitions_tests.module_loader_tests.checks_module.checks_submodule import (
     submodule_check as submodule_check,
 )
